==> 1. Dataset_collection.ipynb <==
import pandas as pd
from huggingface_hub import login

# Log in using your Hugging Face token
login("Use your HuggingFace Token")

import pandas as pd

splits = {'train': 'data/train-00000-of-00001.parquet', 'test': 'data/test-00000-of-00001.parquet'}
df_train = pd.read_parquet("hf://datasets/jayavibhav/prompt-injection-safety/" + splits["train"])
df_test = pd.read_parquet("hf://datasets/jayavibhav/prompt-injection-safety/" + splits["test"])

# Combine the DataFrames
combined_df = pd.concat([df_train, df_test], ignore_index=True)

combined_df

combined_df[combined_df['label'] == 1].to_csv("test_data - Prompt.csv", index=False)
